pilip/fix-encryptor-assert

See merge request Tanker/sdk-native!398

+++ modules/sdk-core/src/Encryptor.cpp
@@ -27,7 +27,7 @@ constexpr bool isHugeClearData(uint64_t dataSize)
 }
 
 template <typename Callable>
-decltype(auto) performEncryptorAction(std::uint8_t version, Callable&& cb)
+decltype(auto) performEncryptorAction(std::uint32_t version, Callable&& cb)
 {
   switch (version)
   {
